+++ TDTChocolate/TDTFoundationAdditions.h
@@ -31,3 +31,4 @@
 #import "FoundationAdditions/NSNumber+TDTFixedWidthAccessors.h"
 #import "FoundationAdditions/NSString+TDTFixedWidthAccessors.h"
 #import "FoundationAdditions/NSData+TDTAESDecryption.h"
+#import "FoundationAdditions/NSURL+TDTQueryParameterEncoding.h"