+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 2.5.?, ??-June-2014
 
   * Fix SHA1 computation for Unicode strings.
+  * Remove dependency between Foundation and Testing additions.
+  * Make dependency between Foundation and Core Data additions explicit.
   * Add method to construct dictionary by removing a key.
   * Add method to filter entries from a dictionary.
   * Add alternative URL Query parameter encoding implementation.