+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   * Add method to determine if one dictionary is a subset of another.
   * Add method to construct dictionary by combining pairs.
   * Add custom XCTAsserts for dates and strings.
+  * Deprecate TDTObjectOrDefault.
 
 2.4.0, 26-June-2014